API: Use hash of outscript instead of address
The usage of a ripe160 for bitcoin addresses in the API and in the
Indexer loses some info, specifically what kind of script it is.
Additionally not all types of scripts fit this mold. At best that means
its not future-proof.

This adds a method to the API in order to select from a Tx the hashed
outscript (thats singlehashed sha256) and refactor the address indexer
to use that instead of the ripe160 address.

The API enums broke a little, so I used the opportunity to break it a
lot and clean up the enums in order to make them more future-proof.
But, yeah, software from before this commit is protocol incompatible
with software after this commit.

+++ indexer/HashStorage.h
@@ -50,11 +50,11 @@ public:
 
     /// Append the hash and return its unique point.
     /// Please note behaviour is not defined if the same hash is appended twice.
-    HashIndexPoint append(const uint160 &hash);
+    HashIndexPoint append(const uint256 &hash);
     /// Find in the collection the hash by point. This is an expensive operation!
-    const uint160 &find(HashIndexPoint point) const;
+    const uint256 &find(HashIndexPoint point) const;
     /// return the point given to the hash.
-    HashIndexPoint lookup(const uint160 &hash) const;
+    HashIndexPoint lookup(const uint256 &hash) const;
 
     /// Flush all caches and make lookup on-disk only
     void finalize();
